+++ backend/apid/actions/events.go
@@ -4,6 +4,7 @@ import (
 	"context"
 
 	"github.com/google/uuid"
+	"github.com/sensu/sensu-go/backend/authentication/jwt"
 	"github.com/sensu/sensu-go/backend/messaging"
 	"github.com/sensu/sensu-go/backend/store"
 
@@ -125,6 +126,12 @@ func (a EventController) CreateOrReplace(ctx context.Context, event *corev2.Even
 			return NewError(InternalErr, err)
 		}
 		event.ID = id[:]
+	}
+
+	if claims := jwt.GetClaimsFromContext(ctx); claims != nil {
+		event.CreatedBy = claims.StandardClaims.Subject
+		event.Check.CreatedBy = claims.StandardClaims.Subject
+		event.Entity.CreatedBy = claims.StandardClaims.Subject
 	}
 
 	// Publish to event pipeline