+++ backend/apid/handlers/create_test.go
@@ -2,14 +2,17 @@ package handlers
 
 import (
 	"bytes"
+	"context"
 	"net/http"
 	"testing"
 
 	"github.com/gorilla/mux"
 	corev2 "github.com/sensu/sensu-go/api/core/v2"
+	"github.com/sensu/sensu-go/backend/authentication/jwt"
 	"github.com/sensu/sensu-go/backend/store"
 	"github.com/sensu/sensu-go/testing/fixture"
 	"github.com/sensu/sensu-go/testing/mockstore"
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 )
 
@@ -94,4 +97,25 @@ func TestHandlers_CreateResource(t *testing.T) {
 			}
 		})
 	}
+}
+
+func TestCreatedByCreate(t *testing.T) {
+	claims, err := jwt.NewClaims(&corev2.User{Username: "admin"})
+	assert.NoError(t, err)
+	ctx := context.WithValue(context.Background(), corev2.ClaimsKey, claims)
+	body := marshal(t, fixture.Resource{ObjectMeta: corev2.ObjectMeta{}})
+
+	store := &mockstore.MockStore{}
+	h := Handlers{
+		Resource: &fixture.Resource{},
+		Store:    store,
+	}
+
+	store.On("CreateResource", mock.Anything, mock.AnythingOfType("*fixture.Resource")).Return(nil)
+
+	req, err := http.NewRequestWithContext(ctx, http.MethodPost, "/", bytes.NewReader(body))
+	assert.NoError(t, err)
+
+	_, err = h.CreateResource(req)
+	assert.NoError(t, err)
 }