+++ backend/eventd/entity.go
@@ -46,6 +46,7 @@ func createProxyEntity(event *corev2.Event, s store.EntityStore) error {
 			entity.Subscriptions = append(entity.Subscriptions, corev2.GetEntitySubscription(entityName))
 		}
 
+		entity.CreatedBy = event.CreatedBy
 		if err := s.UpdateEntity(ctx, entity); err != nil {
 			return err
 		}