+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 
 ### Fixed
 - ensure that check/check config have a non-empty command
+- Check history is now in FIFO order, not ordered by executed timestamp.
 
 ## [5.18.0] - 2020-02-24